+++ packages/stargate/src/accounts.spec.ts
@@ -1,3 +1,4 @@
+import { fromBase64 } from "@cosmjs/encoding";
 import { Any } from "cosmjs-types/google/protobuf/any";
 
 import { accountFromAny } from "./accounts";
@@ -6,10 +7,11 @@ describe("accounts", () => {
   describe("accountFromAny", () => {
     it("works for PeriodicVestingAccount", () => {
       // Queried from chain via `packages/cli/examples/get_akash_vesting_account.ts`.
-      const any = Any.fromJSON({
+      const any = Any.fromPartial({
         typeUrl: "/cosmos.vesting.v1beta1.PeriodicVestingAccount",
-        value:
+        value: fromBase64(
           "CsMBCnoKLGFrYXNoMXF5MHZ1cjNmbDJ1Y3p0cHpjcmZlYTdtYzhqd3o4eGptdnE3cXZ5EkYKHy9jb3Ntb3MuY3J5cHRvLnNlY3AyNTZrMS5QdWJLZXkSIwohA/XsdhwSIKU73TltD9STcaS07FNw0szR4a+oDLr6vikaGDggGxIUCgR1YWt0EgwxNjY2NjY2NzAwMDAaEwoEdWFrdBILMzcxOTAzMzAwMDAiFAoEdWFrdBIMMTY2NjY2NjcwMDAwKOC9wZkGEODvt/sFGhoIgOeEDxITCgR1YWt0Egs4MzMzMzMzNTAwMBoaCIC/ugcSEwoEdWFrdBILNDE2NjY2Njc1MDAaGgiAqMoHEhMKBHVha3QSCzQxNjY2NjY3NTAw",
+        ),
       });
 
       const account = accountFromAny(any);