+++ packages/stargate/src/modules/staking/aminomessages.spec.ts
@@ -18,9 +18,21 @@ import {
   AminoMsgEditValidator,
   AminoMsgUndelegate,
   createStakingAminoConverters,
+  protoDecimalToJson,
 } from "./aminomessages";
 
 describe("AminoTypes", () => {
+  describe("protoDecimalToJson", () => {
+    it("works", () => {
+      expect(protoDecimalToJson("0")).toEqual("0.000000000000000000");
+      expect(protoDecimalToJson("1")).toEqual("0.000000000000000001");
+      expect(protoDecimalToJson("2497")).toEqual("0.000000000000002497");
+      expect(protoDecimalToJson("987000000000000000")).toEqual("0.987000000000000000");
+      expect(protoDecimalToJson("123987000000000000000")).toEqual("123.987000000000000000");
+      expect(protoDecimalToJson("4872000000000000000000")).toEqual("4872.000000000000000000");
+    });
+  });
+
   describe("toAmino", () => {
     it("works for MsgBeginRedelegate", () => {
       const msg: MsgBeginRedelegate = {
@@ -29,7 +41,7 @@ describe("AminoTypes", () => {
         validatorDstAddress: "cosmos1xy4yqngt0nlkdcenxymg8tenrghmek4nmqm28k",
         amount: coin(1234, "ucosm"),
       };
-      const aminoTypes = new AminoTypes(createStakingAminoConverters("cosmos"));
+      const aminoTypes = new AminoTypes(createStakingAminoConverters());
       const aminoMsg = aminoTypes.toAmino({
         typeUrl: "/cosmos.staking.v1beta1.MsgBeginRedelegate",
         value: msg,
@@ -75,7 +87,7 @@ describe("AminoTypes", () => {
         },
         value: coin(1234, "ucosm"),
       };
-      const aminoTypes = new AminoTypes(createStakingAminoConverters("cosmos"));
+      const aminoTypes = new AminoTypes(createStakingAminoConverters());
       const aminoMsg = aminoTypes.toAmino({
         typeUrl: "/cosmos.staking.v1beta1.MsgCreateValidator",
         value: msg,
@@ -114,7 +126,7 @@ describe("AminoTypes", () => {
         validatorAddress: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
         amount: coin(1234, "ucosm"),
       };
-      const aminoTypes = new AminoTypes(createStakingAminoConverters("cosmos"));
+      const aminoTypes = new AminoTypes(createStakingAminoConverters());
       const aminoMsg = aminoTypes.toAmino({
         typeUrl: "/cosmos.staking.v1beta1.MsgDelegate",
         value: msg,
@@ -143,7 +155,7 @@ describe("AminoTypes", () => {
         minSelfDelegation: "123",
         validatorAddress: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
       };
-      const aminoTypes = new AminoTypes(createStakingAminoConverters("cosmos"));
+      const aminoTypes = new AminoTypes(createStakingAminoConverters());
       const aminoMsg = aminoTypes.toAmino({
         typeUrl: "/cosmos.staking.v1beta1.MsgEditValidator",
         value: msg,
@@ -172,7 +184,7 @@ describe("AminoTypes", () => {
         validatorAddress: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
         amount: coin(1234, "ucosm"),
       };
-      const aminoTypes = new AminoTypes(createStakingAminoConverters("cosmos"));
+      const aminoTypes = new AminoTypes(createStakingAminoConverters());
       const aminoMsg = aminoTypes.toAmino({
         typeUrl: "/cosmos.staking.v1beta1.MsgUndelegate",
         value: msg,
@@ -200,7 +212,7 @@ describe("AminoTypes", () => {
           amount: coin(1234, "ucosm"),
         },
       };
-      const msg = new AminoTypes(createStakingAminoConverters("cosmos")).fromAmino(aminoMsg);
+      const msg = new AminoTypes(createStakingAminoConverters()).fromAmino(aminoMsg);
       const expectedValue: MsgBeginRedelegate = {
         delegatorAddress: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
         validatorSrcAddress: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
@@ -239,7 +251,7 @@ describe("AminoTypes", () => {
           value: coin(1234, "ucosm"),
         },
       };
-      const msg = new AminoTypes(createStakingAminoConverters("cosmos")).fromAmino(aminoMsg);
+      const msg = new AminoTypes(createStakingAminoConverters()).fromAmino(aminoMsg);
       const expectedValue: MsgCreateValidator = {
         description: {
           moniker: "validator",
@@ -283,7 +295,7 @@ describe("AminoTypes", () => {
           amount: coin(1234, "ucosm"),
         },
       };
-      const msg = new AminoTypes(createStakingAminoConverters("cosmos")).fromAmino(aminoMsg);
+      const msg = new AminoTypes(createStakingAminoConverters()).fromAmino(aminoMsg);
       const expectedValue: MsgDelegate = {
         delegatorAddress: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
         validatorAddress: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
@@ -311,7 +323,7 @@ describe("AminoTypes", () => {
           validator_address: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
         },
       };
-      const msg = new AminoTypes(createStakingAminoConverters("cosmos")).fromAmino(aminoMsg);
+      const msg = new AminoTypes(createStakingAminoConverters()).fromAmino(aminoMsg);
       const expectedValue: MsgEditValidator = {
         description: {
           moniker: "validator",
@@ -339,7 +351,7 @@ describe("AminoTypes", () => {
           amount: coin(1234, "ucosm"),
         },
       };
-      const msg = new AminoTypes(createStakingAminoConverters("cosmos")).fromAmino(aminoMsg);
+      const msg = new AminoTypes(createStakingAminoConverters()).fromAmino(aminoMsg);
       const expectedValue: MsgUndelegate = {
         delegatorAddress: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
         validatorAddress: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",