+++ packages/stargate/src/queryclient/index.ts
@@ -1,4 +1,4 @@
-export { QueryClient } from "./queryclient";
+export { QueryAbciResponse, QueryClient, QueryStoreResponse } from "./queryclient";
 export {
   createPagination,
   createProtobufRpcClient,