+++ packages/tendermint-rpc/src/tendermint34/adaptor/responses.ts
@@ -732,8 +732,6 @@ interface RpcTxEvent {
   readonly tx: string;
   readonly result: RpcTxData;
   readonly height: string;
-  /** Not set since Tendermint 0.34 */
-  readonly index?: number;
 }
 
 function decodeTxEvent(data: RpcTxEvent): responses.TxEvent {
@@ -743,7 +741,6 @@ function decodeTxEvent(data: RpcTxEvent): responses.TxEvent {
     hash: hashTx(tx),
     result: decodeTxData(data.result),
     height: apiToSmallInt(assertNotEmpty(data.height)),
-    index: may(apiToSmallInt, data.index),
   };
 }